[CI/Build] Move pre-commit only scripts to `tools/pre_commit`
