+++ tools/pre_commit/check_spdx_header.py
