+++ tools/pre_commit/png-lint.sh
