+++ tools/pre_commit/validate_config.py
